Add failure image to test object
The main motivation for this is when there is a test failure, I would like to
easily grab the path for the failed test image. I then take the failed image
and post it to a url for others to see. This simple change allowed me to easily
do that.